Refactor endpoint hooks to own internal file, separate utils as well
